feat: refactor project architecture and fix critical issues
- Fix mixed package manager issue by removing package-lock.json and using Bun exclusively
- Restructure project with proper folder organization:
  * Move layout components from src/pages/layout/ to src/layouts/
  * Rename src/Enum/ to src/types/ for better semantic naming
  * Add new directories: components/, hooks/, services/, utils/
  * Create barrel exports with index.ts files
- Configure DaisyUI in Tailwind config with proper themes
- Update import paths throughout the application
- Add .gitignore entries to prevent npm artifacts
- Validate build and lint processes successfully

BREAKING CHANGES: Import paths have changed from relative to absolute paths using new directory structure
