+++ src/components/index.ts
@@ -1,0 +1,2 @@
+// Export UI components
+// export * from './ui'; // TODO: Add UI components
