+++ src/components/ui/index.ts
@@ -1,0 +1,2 @@
+// UI components will be added here
+export {};
