+++ src/layouts/index.ts
@@ -1,0 +1,3 @@
+export { default as Header } from './Header';
+export { default as Footer } from './Footer';
+export { default as NotifHeader } from './NotifHeader';
