+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+// Utility functions will be added here
+export {};
